requirements.txt
python-telegram-bot==21.4
python-dotenv==1.0.1
websockets==11.0
requests==2.31.0
gql==3.4.0
requests_toolbelt==1.0.0
aiogram==3.1.0
